% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_regmedint_class_user_methods.R
\name{coef.summary_regmedint}
\alias{coef.summary_regmedint}
\title{Extract the result matrix from a summary_regmedint object.}
\usage{
\method{coef}{summary_regmedint}(x, ...)
}
\arguments{
\item{x}{An object with a class of \code{summary_regmedint}.}

\item{...}{For compatibility with the generic.}
}
\value{
A matrix populated with results.
}
\description{
Extract the result matrix from a summary_regmedint object.
}
