% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_calc_myreg_helpers_coef.R
\name{grad_prop_med_yreg_linear}
\alias{grad_prop_med_yreg_linear}
\title{Calculate the gradient of the proportion mediated for yreg linear.}
\usage{
grad_prop_med_yreg_linear(pnde, tnie)
}
\arguments{
\item{pnde}{Pure natural direct effect.}

\item{tnie}{Total natural indirect effect.}
}
\value{
Proportion mediated value.
}
\description{
Calculate the gradient of the proportion mediated for yreg linear case.
}
