% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_regmedint_class_user_methods.R
\name{print.summary_regmedint}
\alias{print.summary_regmedint}
\title{Print method for summary objects from \code{\link{summary.regmedint}}}
\usage{
\method{print}{summary_regmedint}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{summary_regmedint}.}

\item{...}{For compatibility with the generic function.}
}
\value{
Invisibly return the first argument.
}
\description{
Print results contained in a \code{summary_regmedint} object with additional explanation regarding the evaluation settings.
}
