% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_package_information.R
\docType{package}
\name{regmedint-package}
\alias{regmedint-package}
\title{regmedint: A package for regression-based causal mediation analysis}
\description{
The package is an R implementation of regression-based closed-form causal mediation as originally described in Valeri & VanderWeele 2013 and Valeri & VanderWeele 2015 \url{https://www.hsph.harvard.edu/tyler-vanderweele/tools-and-tutorials/}. The earlier version is a sister program of the SAS macro. The current extended version (version 1.0 and later) supports effect modification by covariates (treatment-covariate and mediator-covariate product terms) in mediator and outcome models.
}
\section{Fitting models}{

Use the regmedint function to fit models and set up regression-based causal mediation analysis.
}

\section{Examining results}{

Several methods are available to examine the regmedint object.
print
summary
coef
confint
}

\seealso{
Useful links:
\itemize{
  \item \url{https://kaz-yos.github.io/regmedint/}
  \item Report bugs at \url{https://github.com/kaz-yos/regmedint/issues}
}

}
\author{
\strong{Maintainer}: Yi Li \email{yi.li10@mail.mcgill.ca} (\href{https://orcid.org/0000-0002-9359-210X}{ORCID})

Authors:
\itemize{
  \item Kazuki Yoshida \email{kazukiyoshida@mail.harvard.edu} (\href{https://orcid.org/0000-0002-2030-3549}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Maya Mathur (\href{https://orcid.org/0000-0001-6698-2607}{ORCID}) [contributor]
}

}
\keyword{internal}
