% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_regmedint_class_user_methods.R
\name{summary.regmedint}
\alias{summary.regmedint}
\title{summary method for regmedint object}
\usage{
\method{summary}{regmedint}(
  object,
  a0 = NULL,
  a1 = NULL,
  m_cde = NULL,
  c_cond = NULL,
  args_mreg_fit = list(),
  args_yreg_fit = list(),
  exponentiate = FALSE,
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{An object of the \code{\link{regmedint}} class.}

\item{a0}{A numeric vector of length one.}

\item{a1}{A numeric vector of length one.}

\item{m_cde}{A numeric vector of length one. A mediator value at which the controlled direct effect (CDE) conditional on the adjustment covariates is evaluated. If not provided, the default value supplied to the call to \code{\link{regmedint}} will be used. Only the CDE is affected.}

\item{c_cond}{A numeric vector as long as the number of adjustment covariates. A set of covariate values at which the conditional natural effects are evaluated.}

\item{args_mreg_fit}{A named list of argument to be passed to the method for the \code{mreg_fit} object.}

\item{args_yreg_fit}{A named list of argument to be passed to the method for the \code{mreg_fit} object.}

\item{exponentiate}{Whether to add exponentiated point and confidence limit estimates. When \code{yreg = "linear"}, it is ignored.}

\item{level}{Confidence level for the confidence intervals.}

\item{...}{For compatibility with the generic. Ignored.}
}
\value{
A \code{summary_regmedint} object, which is a list containing the summary objects of the \code{mreg_fit} and the \code{yreg_fit} as well as the mediation analysis results.
}
\description{
Summarize the \code{mreg_fit}, \code{yreg_fit}, and the mediation analysis effect estimates.
}
