% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_package_information.R
\docType{data}
\name{vv2015}
\alias{vv2015}
\title{Example dataset from Valeri and VanderWeele 2015.}
\format{
A tibble with 100 rows and 7 variables:
\describe{
  \item{id}{Positive integer id.}
  \item{x}{Binary treatment assignment variable.}
  \item{m}{Binary mediator variable.}
  \item{y}{Time to event outcome variable.}
  \item{cens}{Binary censoring indicator. Censored is 1.}
  \item{c}{Continuous confounder variable.}
  \item{event}{Binary event indicator. Event is 1.}
}
}
\source{
\url{https://www.hsph.harvard.edu/tyler-vanderweele/tools-and-tutorials/}
}
\usage{
vv2015
}
\description{
An example dataset from Valeri and VanderWeele (2015) <doi:10.1097/EDE.0000000000000253>.
}
\keyword{datasets}
