% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_calc_myreg_mreg_logistic_yreg_linear.R
\name{calc_myreg_mreg_logistic_yreg_linear}
\alias{calc_myreg_mreg_logistic_yreg_linear}
\title{Create calculators for effects and se (mreg logistic / yreg linear)}
\usage{
calc_myreg_mreg_logistic_yreg_linear(
  mreg,
  mreg_fit,
  yreg,
  yreg_fit,
  avar,
  mvar,
  cvar,
  emm_ac_mreg,
  emm_ac_yreg,
  emm_mc_yreg,
  interaction
)
}
\arguments{
\item{mreg}{A character vector of length 1. Mediator regression type: \code{"linear"} or \code{"logistic"}.}

\item{mreg_fit}{Model fit from \code{\link{fit_mreg}}}

\item{yreg}{A character vector of length 1. Outcome regression type: \code{"linear"}, \code{"logistic"}, \code{"loglinear"}, \code{"poisson"}, \code{"negbin"}, \code{"survCox"}, \code{"survAFT_exp"}, or \code{"survAFT_weibull"}.}

\item{yreg_fit}{Model fit from \code{\link{fit_yreg}}}

\item{avar}{A character vector of length 1. Treatment variable name.}

\item{mvar}{A character vector of length 1. Mediator variable name.}

\item{cvar}{A character vector of length > 0. Covariate names. Use \code{NULL} if there is no covariate. However, this is a highly suspicious situation. Even if \code{avar} is randomized, \code{mvar} is not. Thus, there are usually some confounder(s) to account for the common cause structure (confounding) between \code{mvar} and \code{yvar}.}

\item{emm_ac_mreg}{A character vector of length > 0. Effect modifiers names. The covariate vector in treatment-covariate product term in the mediator model.}

\item{emm_ac_yreg}{A character vector of length > 0. Effect modifiers names. The covariate vector in treatment-covariate product term in the outcome model.}

\item{emm_mc_yreg}{A character vector of length > 0. Effect modifiers names. The covariate vector in mediator-covariate product term in outcome model.}

\item{interaction}{A logical vector of length 1. The presence of treatment-mediator interaction in the outcome model. Default to TRUE.}
}
\value{
A list containing a function for effect estimates and a function for corresponding standard errors.
}
\description{
Construct functions for the conditional effect estimates and their standard errors in the mreg logistic / yreg linear setting. Internally, this function deconstructs model objects and feeds parameter estimates to the internal worker functions \code{calc_myreg_mreg_logistic_yreg_linear_est} and \code{calc_myreg_mreg_logistic_yreg_linear_se}.
}
